Fixed up the create_blueprint script
